Added blib support, and made a couple of other changes.
I would like to be able to run the example script from the build
directory. Before this update, the example required ExtUtils::nvcc
to be installed, because the compilation process invokes a fresh
instance of perl which didn't know to include blib. Now the user-level
script checks for a blib invocation and adds -Mblib to the compiler
and linker commands if found.